Simplify trace to events (#449)
## Summary of changes

- Moved `WindowIter` to iterator module.
- Renamed `AssembleFilter`, `EventsFilter`, and `WindowFilter` to
`AssembleIterable`, `EventsIterable`, `WindowIterable`.
- Amalgamated some files in `datatypes` folder.
- Instrumented `produce_eventlist_to_kafka`.
- Removed various `Display` impl blocks.

## Instruction for review/testing

General code review.
Has been tested in simulated pipeline.
