+++ trace-to-events/src/pulse_detection/datatype/trace.rs
@@ -1,9 +1,7 @@
 //! An abstraction of the time-independent types that are processed by the various filters.
-//!
-//! [Todo] This modules can be combined with others for brevity
-use super::Real;
+use super::{Real, Temporal};
 use std::{
-    fmt::{Debug, Display, Formatter, Result},
+    fmt::Debug,
     ops::{Index, IndexMut},
 };
 
@@ -11,9 +9,9 @@ use std::{
 ///
 /// This differs from the TracePoint type in that TracePoint must represent a time value,
 /// whereas TraceValue is time-agnostic.
-pub(crate) trait TraceValue: Default + Clone + Debug + Display {
+pub(crate) trait TraceValue: Default + Clone + Debug {
     /// The type which contains the value of the data point
-    type ContentType: Default + Clone + Debug + Display;
+    type ContentType: Default + Clone + Debug;
 }
 
 impl TraceValue for Real {
@@ -42,19 +40,6 @@ where
 {
     fn default() -> Self {
         Self([T::default(); N])
-    }
-}
-
-impl<const N: usize, T> Display for TraceArray<N, T>
-where
-    T: TraceValue,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let TraceArray(array) = self;
-        for val in array.iter().take(N - 1) {
-            write!(f, "{val},")?;
-        }
-        write!(f, "{0}", array[N - 1])
     }
 }
 
@@ -88,6 +73,7 @@ pub(crate) type RealArray<const N: usize> = TraceArray<N, Real>;
 /// This type allows contains descriptive statistical data.
 #[derive(Default, Clone, Debug)]
 pub(crate) struct Stats {
+    #[allow(unused)] // FIXME
     /// The current value.
     pub(crate) value: Real,
     /// The arithmetic mean.
@@ -97,6 +83,7 @@ pub(crate) struct Stats {
     /// The variance.
     ///
     /// This may have been calculated from applying a window to a range of values.
+    #[allow(unused)] // FIXME
     pub(crate) variance: Real,
 }
 
@@ -110,12 +97,53 @@ impl From<Real> for Stats {
     }
 }
 
-impl Display for Stats {
-    fn fmt(&self, f: &mut Formatter) -> Result {
-        write!(f, "({}, {}, {})", self.value, self.mean, self.variance)
+impl TraceValue for Stats {
+    type ContentType = Stats;
+}
+
+/// Abstracts types that are processed by the various filters.
+///
+/// To implement TracePoint a type must contain time data and a value.
+pub(crate) trait TracePoint: Clone {
+    /// Represents the time of the data point.
+    /// This should be trivially copyable (usually a scalar).
+    type Time: Temporal;
+
+    /// Represents the value of the data point.
+    type Value: TraceValue;
+
+    /// Returns the time of the data point.
+    fn get_time(&self) -> Self::Time;
+
+    /// Returns an immutable reference to the value of the data point.
+    fn get_value(&self) -> &Self::Value;
+
+    /// Take ownership of a clone of the value without destructing the data point.
+    fn clone_value(&self) -> Self::Value {
+        self.get_value().clone()
     }
 }
 
-impl TraceValue for Stats {
-    type ContentType = Stats;
+/// This is the most basic non-trivial TraceData type.
+/// The first element is the TimeType and the second the ValueType.
+/// feedback.
+impl<X, Y> TracePoint for (X, Y)
+where
+    X: Temporal,
+    Y: TraceValue,
+{
+    type Time = X;
+    type Value = Y;
+
+    fn get_time(&self) -> Self::Time {
+        self.0
+    }
+
+    fn get_value(&self) -> &Self::Value {
+        &self.1
+    }
+
+    fn clone_value(&self) -> Self::Value {
+        self.get_value().clone()
+    }
 }
