+++ trace-to-events/src/pulse_detection/iterators/mod.rs
@@ -1,5 +1,9 @@
 //! Provides iterators to convert raw trace data into events and pulses.
+pub(crate) mod assembler;
 pub(crate) mod event;
+pub(crate) mod window;
 
 use super::{Assembler, Detector, Pulse, TracePoint};
-pub(crate) use event::{AssembleFilter, EventFilter};
+pub(crate) use assembler::AssembleIterable;
+pub(crate) use event::EventsIterable;
+pub(crate) use window::WindowIterable;
