+++ trace-to-events/src/pulse_detection/window/smoothing_window.rs
@@ -97,7 +97,7 @@ impl Window for SmoothingWindow {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::pulse_detection::window::WindowFilter;
+    use crate::pulse_detection::iterators::WindowIterable;
     use assert_approx_eq::assert_approx_eq;
 
     #[test]
